requirements.txt
streamlit
pymongo
bcrypt
google-auth
google-auth-oauthlib
google-api-python-client
requests
pandas
plotly
openai       # (for GPT summarization)
python-dotenv
flask